@@ -1,30 +1,31 @@
 package com.eerussianguy.rainbowoaks;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.levelgen.GenerationStep;
+import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.world.BiomeLoadingEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.eventbus.api.IEventBus;
 
-import static com.eerussianguy.rainbowoaks.RainbowOaks.MOD_ID;
-
-@Mod.EventBusSubscriber(bus=Mod.EventBusSubscriber.Bus.FORGE, modid=MOD_ID)
 public class ForgeEventHandler
 {
-    @SubscribeEvent
-    public static void onBiomeLoad(BiomeLoadingEvent event)
+    public static void init()
     {
-        List<ResourceLocation> biomes = ROConfig.COMMON.biomes.get().stream().map(ResourceLocation::new).collect(Collectors.toList());
+        final IEventBus bus = MinecraftForge.EVENT_BUS;
+        bus.addListener(ForgeEventHandler::onBiomeLoad);
+    }
+
+    private static void onBiomeLoad(final BiomeLoadingEvent event)
+    {
+        List<ResourceLocation> biomes = ROConfig.COMMON.biomes.get().stream().map(ResourceLocation::new).toList();
         if (event.getName() != null)
         {
             for (ResourceLocation res : biomes)
             {
                 if (event.getName().compareNamespaced(res) == 0)
                 {
-                    event.getGeneration().getFeatures(GenerationStep.Decoration.VEGETAL_DECORATION).add(() -> ROConfiguredFeatures.RAINBOW_TREES);
+                    event.getGeneration().getFeatures(GenerationStep.Decoration.VEGETAL_DECORATION).add(ROPlacements.RAINBOW_TREES_CHECKED);
                 }
             }
         }
