@@ -1,24 +1,22 @@
 package com.eerussianguy.rainbowoaks;
 
+import java.util.List;
 import java.util.OptionalInt;
-import javax.annotation.Nonnull;
 
-import com.google.common.collect.ImmutableList;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.core.Registry;
-import net.minecraft.data.BuiltinRegistries;
+import net.minecraft.core.Holder;
+import net.minecraft.data.worldgen.features.FeatureUtils;
+import net.minecraft.data.worldgen.placement.TreePlacements;
 import net.minecraft.util.valueproviders.ConstantInt;
-import net.minecraft.world.level.levelgen.feature.stateproviders.SimpleStateProvider;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.levelgen.feature.WeightedPlacedFeature;
+import net.minecraft.world.level.levelgen.feature.stateproviders.BlockStateProvider;
 import net.minecraft.world.level.levelgen.feature.foliageplacers.BlobFoliagePlacer;
 import net.minecraft.world.level.levelgen.feature.foliageplacers.FancyFoliagePlacer;
-import net.minecraft.world.level.levelgen.placement.FrequencyWithExtraChanceDecoratorConfiguration;
-import net.minecraft.world.level.levelgen.placement.FeatureDecorator;
 import net.minecraft.world.level.levelgen.feature.trunkplacers.FancyTrunkPlacer;
 import net.minecraft.world.level.levelgen.feature.trunkplacers.StraightTrunkPlacer;
 
 import static com.eerussianguy.rainbowoaks.RainbowOaks.MOD_ID;
 
-import net.minecraft.data.worldgen.Features;
 import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
 import net.minecraft.world.level.levelgen.feature.Feature;
 import net.minecraft.world.level.levelgen.feature.configurations.FeatureConfiguration;
@@ -28,34 +26,57 @@ import net.minecraft.world.level.levelgen.feature.featuresize.TwoLayersFeatureSi
 
 public class ROConfiguredFeatures
 {
-    public static final ConfiguredFeature<TreeConfiguration, ?> RAINBOW_OAK = register("rainbow_oak",
-        Feature.TREE.configured((new TreeConfiguration.TreeConfigurationBuilder(
-            new SimpleStateProvider(RORegistry.RAINBOW_LOG.get().defaultBlockState()),
-            new StraightTrunkPlacer(4, 2, 0),
-            new SimpleStateProvider(RORegistry.RAINBOW_LEAVES.get().defaultBlockState()),
-            new SimpleStateProvider(RORegistry.RAINBOW_SAPLING.get().defaultBlockState()),
-            new BlobFoliagePlacer(ConstantInt.of(2), ConstantInt.of(0), 3),
-            new TwoLayersFeatureSize(1, 0, 1))).ignoreVines().build()));
+    public static Holder<ConfiguredFeature<TreeConfiguration, ?>> RAINBOW_OAK;
+    public static Holder<ConfiguredFeature<TreeConfiguration, ?>> FANCY_RAINBOW_OAK;
+    public static Holder<ConfiguredFeature<RandomFeatureConfiguration, ?>> RAINBOW_TREES;
 
-    public static final ConfiguredFeature<TreeConfiguration, ?> FANCY_RAINBOW_OAK = register("fancy_rainbow_oak",
-        Feature.TREE.configured((new TreeConfiguration.TreeConfigurationBuilder(
-            new SimpleStateProvider(RORegistry.RAINBOW_LOG.get().defaultBlockState()),
-            new FancyTrunkPlacer(3, 11, 0),
-            new SimpleStateProvider(RORegistry.RAINBOW_LEAVES.get().defaultBlockState()),
-            new SimpleStateProvider(RORegistry.RAINBOW_SAPLING.get().defaultBlockState()),
-            new FancyFoliagePlacer(ConstantInt.of(2), ConstantInt.of(4), 4),
-            new TwoLayersFeatureSize(0, 0, 0, OptionalInt.of(4)))).ignoreVines().build()));
-
-    public static final ConfiguredFeature<?, ?> RAINBOW_TREES = register("rainbow_trees",
-        Feature.RANDOM_SELECTOR.configured(new RandomFeatureConfiguration(
-            ImmutableList.of(FANCY_RAINBOW_OAK.weighted(0.25F)), RAINBOW_OAK))
-            .decorated(Features.Decorators.HEIGHTMAP_SQUARE).decorated(FeatureDecorator.COUNT_EXTRA
-            .configured(new FrequencyWithExtraChanceDecoratorConfiguration(ROConfig.COMMON.extraAttempts.get(), 0.1F, 1)))
-            .rarity(ROConfig.COMMON.rarity.get()));
-
-
-    private static <FC extends FeatureConfiguration> ConfiguredFeature<FC, ?> register(@Nonnull String name, ConfiguredFeature<FC, ?> cf)
+    public static void init()
     {
-        return Registry.register(BuiltinRegistries.CONFIGURED_FEATURE, new ResourceLocation(MOD_ID, name), cf);
+        RAINBOW_OAK = register("rainbow_oak", Feature.TREE, createOak().build());
+        FANCY_RAINBOW_OAK = register("fancy_rainbow_oak", Feature.TREE, createFancyOak().build());
+    }
+
+    public static void init2()
+    {
+        RAINBOW_TREES = register("rainbow_trees", Feature.RANDOM_SELECTOR,
+            new RandomFeatureConfiguration(List.of(
+                new WeightedPlacedFeature(ROPlacements.RAINBOW_OAK_CHECKED, 0.2F),
+                new WeightedPlacedFeature(ROPlacements.FANCY_RAINBOW_OAK_CHECKED, 0.1F)
+            ),
+                TreePlacements.OAK_BEES_002
+            )
+        );
+    }
+
+    private static TreeConfiguration.TreeConfigurationBuilder createFancyOak()
+    {
+        return (new TreeConfiguration.TreeConfigurationBuilder(
+            BlockStateProvider.simple(RORegistry.RAINBOW_LOG.get()),
+            new FancyTrunkPlacer(3, 11, 0),
+            BlockStateProvider.simple(RORegistry.RAINBOW_LEAVES.get()),
+            new FancyFoliagePlacer(ConstantInt.of(2), ConstantInt.of(4), 4),
+            new TwoLayersFeatureSize(0, 0, 0,OptionalInt.of(4))))
+            .ignoreVines();
+    }
+
+    private static TreeConfiguration.TreeConfigurationBuilder createOak()
+    {
+        return createStraightBlobTree(RORegistry.RAINBOW_LOG.get(), RORegistry.RAINBOW_LEAVES.get(), 4, 2, 0, 2).ignoreVines();
+    }
+
+    private static TreeConfiguration.TreeConfigurationBuilder createStraightBlobTree(Block log, Block leaves, int baseHeight, int heightRandA, int heightRandB, int offset)
+    {
+        return new TreeConfiguration.TreeConfigurationBuilder(
+            BlockStateProvider.simple(log),
+            new StraightTrunkPlacer(baseHeight, heightRandA, heightRandB),
+            BlockStateProvider.simple(leaves),
+            new BlobFoliagePlacer(ConstantInt.of(offset),ConstantInt.of(0), 3),
+            new TwoLayersFeatureSize(1, 0, 1)
+        );
+    }
+
+    private static <FC extends FeatureConfiguration, F extends Feature<FC>> Holder<ConfiguredFeature<FC, ?>> register(String name, F feature, FC config)
+    {
+        return FeatureUtils.register(MOD_ID + ":" + name, feature, config);
     }
 }
