@@ -2,46 +2,45 @@ package com.eerussianguy.rainbowoaks;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import net.minecraft.client.renderer.RenderType;
-import net.minecraft.client.renderer.ItemBlockRenderTypes;
-import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
+import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.fml.loading.FMLEnvironment;
 
 import static com.eerussianguy.rainbowoaks.RainbowOaks.MOD_ID;
 
-// The value here should match an entry in the META-INF/mods.toml file
 @Mod(MOD_ID)
 public class RainbowOaks
 {
-    // Directly reference a log4j logger.
     private static final Logger LOGGER = LogManager.getLogger();
-
-    public static Logger getLogger()
-    {
-        return LOGGER;
-    }
 
     public static final String MOD_ID = "rainbowoaks";
 
     public RainbowOaks()
     {
-        IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
-        modEventBus.addListener(this::clientSetup);
+        IEventBus modBus = FMLJavaModLoadingContext.get().getModEventBus();
+        modBus.addListener(this::commonSetup);
 
-        // Register ourselves for server and other game events we are interested in
-        MinecraftForge.EVENT_BUS.register(this);
+        RORegistry.BLOCKS.register(modBus);
+        RORegistry.ITEMS.register(modBus);
 
-        RORegistry.BLOCKS.register(modEventBus);
-        RORegistry.ITEMS.register(modEventBus);
-
+        if (FMLEnvironment.dist == Dist.CLIENT)
+        {
+            ClientEventHandler.init();
+        }
+        ForgeEventHandler.init();
         ROConfig.init();
     }
 
-    private void clientSetup(final FMLClientSetupEvent event)
+    private void commonSetup(final FMLCommonSetupEvent event)
     {
-        ItemBlockRenderTypes.setRenderLayer(RORegistry.RAINBOW_SAPLING.get(), RenderType.cutout());
+        event.enqueueWork(() -> {
+            ROConfiguredFeatures.init();
+            ROPlacements.init();
+            ROConfiguredFeatures.init2();
+            ROPlacements.init2();
+        });
     }
 }
